Changed files, refactoring

+++ src/Timer.hpp
@@ -18,7 +18,7 @@ public:
     void pause() {
         if (is_running) {
             auto current_time = std::chrono::high_resolution_clock::now();
-            accumulated_time += std::chrono::duration<double>(current_time - start_time).count();
+            accumulated_time += std::chrono::duration<float>(current_time - start_time).count();
             is_running = false;
         }
     }
@@ -30,27 +30,27 @@ public:
         }
     }
 
-    void add_seconds(double seconds) {
-        if (is_running) 
+    void add_seconds(float seconds) {
+        if (is_running)
         {
             // Якщо таймер працює, оновлюємо accumulated_time з урахуванням поточного часу
             auto current_time = std::chrono::high_resolution_clock::now();
-            accumulated_time += std::chrono::duration<double>(current_time - start_time).count();
+            accumulated_time += std::chrono::duration<float>(current_time - start_time).count();
             start_time = current_time; // Оновлюємо start_time
         }
         accumulated_time += seconds; // Додаємо задані секунди
     }
 
-    double elapsedSeconds() const {
+    float elapsedSeconds() const {
         if (is_running) {
             auto current_time = std::chrono::high_resolution_clock::now();
-            return accumulated_time + std::chrono::duration<double>(current_time - start_time).count();
+            return accumulated_time + std::chrono::duration<float>(current_time - start_time).count();
         } else {
             return accumulated_time;
         }
     }
 
-    double elapsedMinutes() const {
+    float elapsedMinutes() const {
         return elapsedSeconds() / 60.0;
     }
 
@@ -58,7 +58,7 @@ public:
         if (first_run) {
             return "00:00";
         }
-        double seconds = elapsedSeconds();
+        float seconds = elapsedSeconds();
         int minutes = static_cast<int>(seconds / 60);
         int seconds_d = static_cast<int>(seconds) % 60;
 
@@ -67,7 +67,7 @@ public:
         {
             result = "0" + std::to_string(minutes) + ":";
         }
-        else 
+        else
         {
             result = std::to_string(minutes) + ":";
         }
@@ -76,15 +76,15 @@ public:
         {
             result += "0" + std::to_string(seconds_d);
         }
-        else 
+        else
         {
             result += std::to_string(seconds_d);
         }
-        
+
         return result; //std::to_string(minutes) + ":" + std::to_string(seconds_d);
     }
 private:
     std::chrono::time_point<std::chrono::high_resolution_clock> start_time;
-    double accumulated_time = 0; // Час, накопичений перед паузою
+    float accumulated_time = 0; // Час, накопичений перед паузою
     bool is_running = true; // Стан таймера: працює чи на паузі
 };
